app-requirements.txt
# Flask app dependencies only
# (excludes training/data tools: dvc, kaggle, pyyaml)

flask
gunicorn
pandas
numpy
scikit-learn
mlflow
dagshub
python-dotenv
pyrebase4
urllib3<2.0.0
